Fixed issue wherein the CalibrationQueueValue's did not write intended values.
Refactored some CalibrationQueueValues's variables to make more sense given context.

+++ library/src/main/java/com/tenkiv/tekdaqc/hardware/CommandBuilder.java
@@ -509,7 +509,7 @@ public final class CommandBuilder {
         return new CalibrationQueueValue(
                 (byte) Commands.WRITE_GAIN_CALIBRATION_VALUE.ordinal(),
                 temp,
-                (int) value,
+                value,
                 (byte) buffer.ordinal(),
                 (byte) rate.ordinal(),
                 (byte) gain.ordinal(),
